split entities into files
