+++ webknossos/webknossos/skeleton/__init__.py
@@ -1,6 +1,9 @@
 from os import PathLike
 
-from webknossos.skeleton.entities import Graph, Group, Node, Skeleton
+from webknossos.skeleton.graph import Graph
+from webknossos.skeleton.group import Group
+from webknossos.skeleton.node import Node
+from webknossos.skeleton.skeleton import Skeleton
 
 
 def open_nml(file_path: PathLike) -> Skeleton:
